requirements.txt
django==1.8
django-mptt
djangorestframework
